+++ .gitignore
@@ -62,3 +62,5 @@ amplifytools.xcconfig
 # amplify
 .amplify
 amplifyconfiguration*
+
+certificates
