+++ src/app/(home)/home/loading.js
@@ -1,0 +1,7 @@
+export default function HomeLoading() {
+    return (
+        <div>
+            <h1>Loading...</h1>
+        </div>
+    )
+}
