+++ src/app/layout.js
@@ -27,7 +27,8 @@ const MainApp = ({ children }) => {
             <Authenticator.Provider>
                 <View>
                     <CssBaseline/>
-                    <Box>
+                    <Box
+                    >
                         {children}
                     </Box>
                 </View>
@@ -44,7 +45,9 @@ export default function RootLayout({ children }) {
             <head>
                 <link rel="icon" href="/favicon.png" />
             </head>
-            <body>
+            <body
+
+            >
                 <AppRouterCacheProvider>
                     <ConfigureAmplifyClientSide />
                     <ThemeContextProvider>
