+++ src/components/Pages/MobileLanding.js
@@ -1,0 +1,7 @@
+export default function MobileLanding() {
+    return (
+        <div>
+            <h1>Mobile Landing Page</h1>
+        </div>
+    )
+}
